Move collection of data from contactForm to api NewsletterBrevo function.
